Add files via upload
IT FINALLY WORKS! Shout-out to Mike Hector and Kevin Green for helping me with the concept of gain tuning for this kind of system. Next is to read the "Spring-Legged Locomotion on Uneven Ground: A Control Approach to Keep the Running Speed Constant" paper suggested by Kevin in order to make a more robust controller for this model.